refactor: move getScreenTitle function to Utils and update NavigationHost import
- Moved getScreenTitle function to Utils.kt for better code organization.
- Updated NavigationHost to use the new location for the getScreenTitle function.
