+++ app/src/main/java/com/pgmv/bandify/utils/Utils.kt
@@ -40,3 +40,16 @@ fun formatTimeDifference(createdAt: String): String {
         else -> createdDateTime.format(DateTimeFormatter.ofPattern("dd/MM/yyyy"))
     }
 }
+
+fun getScreenTitle(route: String): String {
+    return when (route) {
+        "home" -> "Bandify"
+        "agenda" -> "Agenda"
+        "repertorio?event_id={eventId}" -> "Repertório"
+        "arquivos" -> "Arquivos"
+        "perfil" -> "Perfil"
+        "novo_evento" -> "Adicionar Evento"
+        "nova_musica" -> "Adicionar Música"
+        else -> "Bandify"
+    }
+}
